allow custom elements for picture link
Allow `<PerseidsHeader>` to receive `component` and `props` props. If
`component` is given, it will use that component instead of `<a>` for
the component wrapping the image in the header. If `props` is given,
those will be the props given to the component, instead of the default
`{ href: 'https://www.perseids.org' }`.